requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Deep Learning
tensorflow>=2.8.0
keras>=2.8.0

# NLP Libraries
nltk>=3.6.0
gensim>=4.1.0
spacy>=3.2.0

# Text Processing
regex>=2021.0.0
textblob>=0.15.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Web Framework (for API deployment)
flask>=2.0.0
flask-cors>=3.0.10

# Configuration
pyyaml>=5.4.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
python-dateutil>=2.8.0

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.4.0
ipywidgets>=7.6.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# Additional Tools
openpyxl>=3.0.9  # For Excel file support
xlrd>=2.0.1  # For Excel file reading
requests>=2.26.0
